==> 06 Generate aggregate table.R <==
# GENERATE DATABASE AGGREGATE TABLE
# 2 sec

# Packages
library(sqldf)
library(openxlsx)

# Clear memory and start stopwatch
rm(list = objects())
t0 = Sys.time()

# Parameters
path.database = "../Data/6 Database/"
obs.table.csv = "SIS Peru obs.table.csv"
agr.sector.csv = "SIS Peru agr.sector.csv"
agr.sector.xlsx = "SIS Peru agr.sector.xlsx"

# start stopwatch
t0 = Sys.time()

# 1. Load data
df_data = read.csv2(paste0(path.database, obs.table.csv))

# 2. Create sql group by query
sqltxt = "select [Sector.Cod], [Sector.Nombre.Corto], [Sector.Cod.Nombre], "
# Code generated in Code Generator.xlsx spreadsheet
sqltxt = paste0(sqltxt, "count(*) as [Cantidad.Firmas], ")
sqltxt = paste0(sqltxt, "sum([Gasto.ID.Interno]) as [Gasto.ID.Interno], ")
sqltxt = paste0(sqltxt, "sum([Gasto.ID.Externo]) as [Gasto.ID.Externo], ")
sqltxt = paste0(sqltxt, "sum([Gasto.ID.Total]) as [Gasto.ID.Total], ")
sqltxt = paste0(sqltxt, "sum([Gasto.Innovacion]) as [Gasto.Innovacion], ")
sqltxt = paste0(sqltxt, "sum([Gasto.ID.Interno]) / cast(sum([Ventas]) as float) as [Intensidad.ID.Interno], ")
sqltxt = paste0(sqltxt, "sum([Gasto.ID.Externo]) / cast(sum([Ventas]) as float) as [Intensidad.ID.Externo], ")
sqltxt = paste0(sqltxt, "sum([Gasto.ID.Total]) / cast(sum([Ventas]) as float) as [Intensidad.ID], ")
sqltxt = paste0(sqltxt, "sum([Gasto.Innovacion]) / cast(sum([Ventas]) as float) as [Intensidad.Innovacion], ")
sqltxt = paste0(sqltxt, "sum([Gasto.ID.Interno]) / cast(sum([Gasto.Innovacion]) as float) as [Pct.ID.Interno.Innovacion], ")
sqltxt = paste0(sqltxt, "sum([Gasto.ID.Externo]) / cast(sum([Gasto.Innovacion]) as float) as [Pct.ID.Externo.Innovacion], ")
sqltxt = paste0(sqltxt, "sum([Gasto.ID.Total]) / cast(sum([Gasto.Innovacion]) as float) as [Pct.ID.Total.Innovacion], ")
sqltxt = paste0(sqltxt, "sum([Firma.Innovadora]*[Gasto.ID.Interno]) as [Gasto.ID.Interno.FI], ")
sqltxt = paste0(sqltxt, "sum([Firma.Innovadora]*[Gasto.ID.Externo]) as [Gasto.ID.Externo.FI], ")
sqltxt = paste0(sqltxt, "sum([Firma.Innovadora]*[Gasto.ID.Total]) as [Gasto.ID.Total.FI], ")
sqltxt = paste0(sqltxt, "sum([Firma.Innovadora]*[Gasto.Innovacion]) as [Gasto.Innovacion.FI], ")
sqltxt = paste0(sqltxt, "sum([Firma.Innovadora]*[Gasto.ID.Interno]) / cast(sum([Firma.Innovadora]*[Ventas]) as float) as [Intensidad.ID.Interno.FI], ")
sqltxt = paste0(sqltxt, "sum([Firma.Innovadora]*[Gasto.ID.Externo]) / cast(sum([Firma.Innovadora]*[Ventas]) as float) as [Intensidad.ID.Externo.FI], ")
sqltxt = paste0(sqltxt, "sum([Firma.Innovadora]*[Gasto.ID.Total]) / cast(sum([Firma.Innovadora]*[Ventas]) as float) as [Intensidad.ID.FI], ")
sqltxt = paste0(sqltxt, "sum([Firma.Innovadora]*[Gasto.Innovacion]) / cast(sum([Firma.Innovadora]*[Ventas]) as float) as [Intensidad.Innovacion.FI], ")
sqltxt = paste0(sqltxt, "sum([Firma.Innovadora]*[Gasto.ID.Interno]) / cast(sum([Firma.Innovadora]*[Gasto.Innovacion]) as float) as [Pct.ID.Interno.Innovacion.FI], ")
sqltxt = paste0(sqltxt, "sum([Firma.Innovadora]*[Gasto.ID.Externo]) / cast(sum([Firma.Innovadora]*[Gasto.Innovacion]) as float) as [Pct.ID.Externo.Innovacion.FI], ")
sqltxt = paste0(sqltxt, "sum([Firma.Innovadora]*[Gasto.ID.Total]) / cast(sum([Firma.Innovadora]*[Gasto.Innovacion]) as float) as [Pct.ID.Total.Innovacion.FI], ")
sqltxt = paste0(sqltxt, "sum([Financiamiento.Propio]) / cast(count(*) as float) as [Pct.Financiamiento.Propio], ")
sqltxt = paste0(sqltxt, "sum([Firma.Innovadora]*[Financiamiento.Propio]) / cast(sum([Firma.Innovadora]) as float) as [Pct.Financiamiento.Propio.FI], ")
sqltxt = paste0(sqltxt, "sum([Financiamiento.Externo]) / cast(count(*) as float) as [Pct.Financiamiento.Externo], ")
sqltxt = paste0(sqltxt, "sum([Firma.Innovadora]*[Financiamiento.Externo]) / cast(sum([Firma.Innovadora]) as float) as [Pct.Financiamiento.Externo.FI], ")
sqltxt = paste0(sqltxt, "sum([RRHH.Universitaria]) / cast(sum([RRHH]) as float) as [Pct.RRHH.Universitaria], ")
sqltxt = paste0(sqltxt, "sum([Firma.Innovadora]*[RRHH.Universitaria]) / cast(sum([Firma.Innovadora]*[RRHH]) as float) as [Pct.RRHH.Universitaria.FI], ")
sqltxt = paste0(sqltxt, "sum([RRHH.ID]) / cast(sum([RRHH]) as float) as [Pct.RRHH.ID], ")
sqltxt = paste0(sqltxt, "sum([Firma.Innovadora]*[RRHH.ID]) / cast(sum([Firma.Innovadora]*[RRHH]) as float) as [Pct.RRHH.ID.FI], ")
sqltxt = paste0(sqltxt, "sum([Firma.Innovadora]) as [Cantidad.Firmas.Innovadoras], ")
sqltxt = paste0(sqltxt, "sum([Innovacion.Producto]) as [Innovacion.Producto], ")
sqltxt = paste0(sqltxt, "sum([Innovacion.Proceso]) as [Innovacion.Proceso], ")
sqltxt = paste0(sqltxt, "sum([Innovacion.Organizacional]) as [Innovacion.Organizacional], ")
sqltxt = paste0(sqltxt, "sum([Innovacion.Marketing]) as [Innovacion.Marketing], ")
sqltxt = paste0(sqltxt, "avg([Grado.Innovacion]) as [Grado.Innovacion], ")
sqltxt = paste0(sqltxt, "sum([Firma.Innovadora]) / cast(count(*) as float) as [Pct.Firmas.Innovadoras], ")
sqltxt = paste0(sqltxt, "sum([Innovacion.Producto]) / cast(count(*) as float) as [Pct.Innovacion.Producto], ")
sqltxt = paste0(sqltxt, "sum([Innovacion.Proceso]) / cast(count(*) as float) as [Pct.Innovacion.Proceso], ")
sqltxt = paste0(sqltxt, "sum([Innovacion.Organizacional]) / cast(count(*) as float) as [Pct.Innovacion.Organizacional], ")
sqltxt = paste0(sqltxt, "sum([Innovacion.Marketing]) / cast(count(*) as float) as [Pct.Innovacion.Marketing], ")
sqltxt = paste0(sqltxt, "sum([Cooperacion.Firmas]) as [Cooperacion.Firmas], ")
sqltxt = paste0(sqltxt, "sum([Cooperacion.Firmas]) / cast(count(*) as float) as [Pct.Cooperacion.Firmas], ")
sqltxt = paste0(sqltxt, "sum([Firma.Innovadora]*[Innovacion.Producto]) as [Innovacion.Producto.FI], ")
sqltxt = paste0(sqltxt, "sum([Firma.Innovadora]*[Innovacion.Proceso]) as [Innovacion.Proceso.FI], ")
sqltxt = paste0(sqltxt, "sum([Firma.Innovadora]*[Innovacion.Organizacional]) as [Innovacion.Organizacional.FI], ")
sqltxt = paste0(sqltxt, "sum([Firma.Innovadora]*[Innovacion.Marketing]) as [Innovacion.Marketing.FI], ")
sqltxt = paste0(sqltxt, "sum([Firma.Innovadora]*[Innovacion.Producto]) / cast(sum([Firma.Innovadora]) as float) as [Pct.Innovacion.Producto.FI], ")
sqltxt = paste0(sqltxt, "sum([Firma.Innovadora]*[Innovacion.Proceso]) / cast(sum([Firma.Innovadora]) as float) as [Pct.Innovacion.Proceso.FI], ")
sqltxt = paste0(sqltxt, "sum([Firma.Innovadora]*[Innovacion.Organizacional]) / cast(sum([Firma.Innovadora]) as float) as [Pct.Innovacion.Organizacional.FI], ")
sqltxt = paste0(sqltxt, "sum([Firma.Innovadora]*[Innovacion.Marketing]) / cast(sum([Firma.Innovadora]) as float) as [Pct.Innovacion.Marketing.FI], ")
sqltxt = paste0(sqltxt, "sum([Firma.Innovadora]*[Cooperacion.Firmas]) as [Cooperacion.Firmas.FI], ")
sqltxt = paste0(sqltxt, "sum([Firma.Innovadora]*[Cooperacion.Firmas]) / cast(sum([Firma.Innovadora]) as float) as [Pct.Cooperacion.Firmas.FI], ")
sqltxt = paste0(sqltxt, "sum([Propiedad.Industrial]) as [Propiedad.Industrial], ")
sqltxt = paste0(sqltxt, "sum([Marca]) as [Marca], ")
sqltxt = paste0(sqltxt, "sum([Patente]) as [Patente], ")
sqltxt = paste0(sqltxt, "sum([Modelos.Utilidad]) as [Modelos.Utilidad], ")
sqltxt = paste0(sqltxt, "sum([Propiedad.Industrial]) / cast(count(*) as float) as [Pct.Propiedad.Industrial], ")
sqltxt = paste0(sqltxt, "sum([Marca]) / cast(count(*) as float) as [Pct.Marca], ")
sqltxt = paste0(sqltxt, "sum([Patente]) / cast(count(*) as float) as [Pct.Patente], ")
sqltxt = paste0(sqltxt, "sum([Modelos.Utilidad]) / cast(count(*) as float) as [Pct.Modelos.Utilidad], ")
sqltxt = paste0(sqltxt, "sum([Firma.Innovadora]*[Propiedad.Industrial]) as [Propiedad.Industrial.FI], ")
sqltxt = paste0(sqltxt, "sum([Firma.Innovadora]*[Marca]) as [Marca.FI], ")
sqltxt = paste0(sqltxt, "sum([Firma.Innovadora]*[Patente]) as [Patente.FI], ")
sqltxt = paste0(sqltxt, "sum([Firma.Innovadora]*[Modelos.Utilidad]) as [Modelos.Utilidad.FI], ")
sqltxt = paste0(sqltxt, "sum([Firma.Innovadora]*[Propiedad.Industrial]) / cast(sum([Firma.Innovadora]) as float) as [Pct.Propiedad.Industrial.FI], ")
sqltxt = paste0(sqltxt, "sum([Firma.Innovadora]*[Marca]) / cast(sum([Firma.Innovadora]) as float) as [Pct.Marca.FI], ")
sqltxt = paste0(sqltxt, "sum([Firma.Innovadora]*[Patente]) / cast(sum([Firma.Innovadora]) as float) as [Pct.Patente.FI], ")
sqltxt = paste0(sqltxt, "sum([Firma.Innovadora]*[Modelos.Utilidad]) / cast(sum([Firma.Innovadora]) as float) as [Pct.Modelos.Utilidad.FI], ")
sqltxt = paste0(sqltxt, "sum([Informacion.Interna]) as [Informacion.Interna], ")
sqltxt = paste0(sqltxt, "sum([Informacion.Proveedores]) as [Informacion.Proveedores], ")
sqltxt = paste0(sqltxt, "sum([Informacion.Clientes]) as [Informacion.Clientes], ")
sqltxt = paste0(sqltxt, "sum([Informacion.Firmas]) as [Informacion.Firmas], ")
sqltxt = paste0(sqltxt, "sum([Informacion.Consultores]) as [Informacion.Consultores], ")
sqltxt = paste0(sqltxt, "sum([Informacion.Universidades]) as [Informacion.Universidades], ")
sqltxt = paste0(sqltxt, "sum([Informacion.Institutos.Investigacion]) as [Informacion.Institutos.Investigacion], ")
sqltxt = paste0(sqltxt, "sum([Informacion.Publicaciones]) as [Informacion.Publicaciones], ")
sqltxt = paste0(sqltxt, "sum([Informacion.Otros]) as [Informacion.Otros], ")
sqltxt = paste0(sqltxt, "sum([Informacion.Interna]) / cast(count(*) as float) as [Pct.Informacion.Interna], ")
sqltxt = paste0(sqltxt, "sum([Informacion.Proveedores]) / cast(count(*) as float) as [Pct.Informacion.Proveedores], ")
sqltxt = paste0(sqltxt, "sum([Informacion.Clientes]) / cast(count(*) as float) as [Pct.Informacion.Clientes], ")
sqltxt = paste0(sqltxt, "sum([Informacion.Firmas]) / cast(count(*) as float) as [Pct.Informacion.Firmas], ")
sqltxt = paste0(sqltxt, "sum([Informacion.Consultores]) / cast(count(*) as float) as [Pct.Informacion.Consultores], ")
sqltxt = paste0(sqltxt, "sum([Informacion.Universidades]) / cast(count(*) as float) as [Pct.Informacion.Universidades], ")
sqltxt = paste0(sqltxt, "sum([Informacion.Institutos.Investigacion]) / cast(count(*) as float) as [Pct.Informacion.Institutos.Investigacion], ")
sqltxt = paste0(sqltxt, "sum([Informacion.Publicaciones]) / cast(count(*) as float) as [Pct.Informacion.Publicaciones], ")
sqltxt = paste0(sqltxt, "sum([Informacion.Otros]) / cast(count(*) as float) as [Pct.Informacion.Otros], ")
sqltxt = paste0(sqltxt, "sum([Firma.Innovadora]*[Informacion.Interna]) as [Informacion.Interna.FI], ")
sqltxt = paste0(sqltxt, "sum([Firma.Innovadora]*[Informacion.Proveedores]) as [Informacion.Proveedores.FI], ")
sqltxt = paste0(sqltxt, "sum([Firma.Innovadora]*[Informacion.Clientes]) as [Informacion.Clientes.FI], ")
sqltxt = paste0(sqltxt, "sum([Firma.Innovadora]*[Informacion.Firmas]) as [Informacion.Firmas.FI], ")
sqltxt = paste0(sqltxt, "sum([Firma.Innovadora]*[Informacion.Consultores]) as [Informacion.Consultores.FI], ")
sqltxt = paste0(sqltxt, "sum([Firma.Innovadora]*[Informacion.Universidades]) as [Informacion.Universidades.FI], ")
sqltxt = paste0(sqltxt, "sum([Firma.Innovadora]*[Informacion.Institutos.Investigacion]) as [Informacion.Institutos.Investigacion.FI], ")
sqltxt = paste0(sqltxt, "sum([Firma.Innovadora]*[Informacion.Publicaciones]) as [Informacion.Publicaciones.FI], ")
sqltxt = paste0(sqltxt, "sum([Firma.Innovadora]*[Informacion.Otros]) as [Informacion.Otros.FI], ")
sqltxt = paste0(sqltxt, "sum([Firma.Innovadora]*[Informacion.Interna]) / cast(sum([Firma.Innovadora]) as float) as [Pct.Informacion.Interna.FI], ")
sqltxt = paste0(sqltxt, "sum([Firma.Innovadora]*[Informacion.Proveedores]) / cast(sum([Firma.Innovadora]) as float) as [Pct.Informacion.Proveedores.FI], ")
sqltxt = paste0(sqltxt, "sum([Firma.Innovadora]*[Informacion.Clientes]) / cast(sum([Firma.Innovadora]) as float) as [Pct.Informacion.Clientes.FI], ")
sqltxt = paste0(sqltxt, "sum([Firma.Innovadora]*[Informacion.Firmas]) / cast(sum([Firma.Innovadora]) as float) as [Pct.Informacion.Firmas.FI], ")
sqltxt = paste0(sqltxt, "sum([Firma.Innovadora]*[Informacion.Consultores]) / cast(sum([Firma.Innovadora]) as float) as [Pct.Informacion.Consultores.FI], ")
sqltxt = paste0(sqltxt, "sum([Firma.Innovadora]*[Informacion.Universidades]) / cast(sum([Firma.Innovadora]) as float) as [Pct.Informacion.Universidades.FI], ")
sqltxt = paste0(sqltxt, "sum([Firma.Innovadora]*[Informacion.Institutos.Investigacion]) / cast(sum([Firma.Innovadora]) as float) as [Pct.Informacion.Institutos.Investigacion.FI], ")
sqltxt = paste0(sqltxt, "sum([Firma.Innovadora]*[Informacion.Publicaciones]) / cast(sum([Firma.Innovadora]) as float) as [Pct.Informacion.Publicaciones.FI], ")
sqltxt = paste0(sqltxt, "sum([Firma.Innovadora]*[Informacion.Otros]) / cast(sum([Firma.Innovadora]) as float) as [Pct.Informacion.Otros.FI], ")
sqltxt = paste0(sqltxt, "sum([Ventas]) as [Ventas], ")
sqltxt = paste0(sqltxt, "sum([Ventas]) / cast(count(*) as float) as [Ventas.Media], ")
sqltxt = paste0(sqltxt, "sum([Firma.Innovadora]*[Ventas]) as [Ventas.FI], ")
sqltxt = paste0(sqltxt, "sum([Firma.Innovadora]*[Ventas]) / cast(sum([Firma.Innovadora]) as float) as [Ventas.Media.FI], ")
# End of code generated in Code Generator.xlsx spreadsheet
sqltxt = substr(sqltxt, 1, nchar(sqltxt)-2) # Remove comma at the end
sqltxt = paste(sqltxt, "from df_data group by [Sector.Cod], [Sector.Nombre.Corto], [Sector.Cod.Nombre] order by [Sector.Cod]")
df_sector = sqldf(sqltxt)

# 3. Save to database: csv and xlsx
write.csv2(df_sector, file = paste0(path.database, agr.sector.csv), row.names = F, na = "")
df_sector_xlsx = df_sector
names(df_sector_xlsx) = gsub(".", " ", names(df_sector_xlsx), fixed = T)
hs = createStyle(textDecoration = "bold", fgFill = "#DDEBF7", wrapText = T)
write.xlsx(df_sector_xlsx, file = paste0(path.database, agr.sector.xlsx), headerStyle = hs, firstRow = T, colWidths = 14)

# Show time taken
print(Sys.time() - t0)
